# app/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExpenseWise REST API",
    "version": "1.0.0",
    "description": "Developer API for ExpenseWise personal finance tracker. All endpoints return JSON and require authorization headers for private data."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token",
        "description": "Enter your personal secure API Token generated from settings."
      }
    }
  },
  "paths": {
...
    "/v1/analytics/summary": {
      "get": {
        "tags": ["Reports & Analytics"],
        "summary": "Fetch summary metrics and category allocations",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "start_date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "end_date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Overview aggregations.",
            "content": {
              "application/json": {
                "example": {
                  "metrics": {
                    "total_spending": 3200.45,
                    "daily_average": 106.68,
                    "total_count": 10
                  },
                  "categories": [
                    { "category": "Food", "total": 120.0, "pct": 10.0, "count": 2 }
                  ],
                  "default_currency": "INR"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analytics/trends": {
      "get": {
        "tags": ["Reports & Analytics"],
        "summary": "Fetch category-wise sums and trends",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "start_date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "end_date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Category allocations and trends.",
            "content": {
              "application/json": {
                "example": {
                  "category_distribution": {
                    "Food": 320.65,
                    "Travel": 165.08
                  },
                  "default_currency": "INR"
                }
              }
            }
          }
        }
      }
    },
...
  }
}